test: share the sync config seeding; document the fail-open dependency (issue #2496)
Round-1 review nitpicks (no blocking findings):

- The new rows skipped the config seeding its sibling SyncFeedPassDeferralTest
  performs, re-triggering 12 pre-existing 'Undefined array key' sites inside
  pfb_global() and adding 33 trigger listings to the suite's warning detail
  (total count unchanged; nothing failed). Rather than duplicate 40 lines, the
  helper is extracted to tests/php/SyncPrereqSeedTrait.php following the repo's
  require_once-trait convention (PfbNoPhpWarningTrait), and BOTH tests use it.

- The stage/publish row's reachability depends on pfb_feed_pass_acquire()
  FAILING OPEN when its lock fopen() fails (pfblockerng.inc:17685) — it logs
  'proceeding unlocked' and returns TRUE, so the feed-pass guard, which runs
  first, does not abort the sync. The review verified this empirically. Now
  documented at the row, including that a flip to fail-closed turns the row RED
  rather than passing vacuously.

Full suite unchanged: 5325 tests, 0 failures, 29 warnings (same as before this
PR). phpstan [OK] No errors; phpcs clean.
