+++ tests/php/SyncFeedPassDeferralTest.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+require_once __DIR__ . '/SyncPrereqSeedTrait.php';
+
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -22,6 +24,8 @@ use PHPUnit\Framework\TestCase;
  */
 final class SyncFeedPassDeferralTest extends TestCase
 {
+	use SyncPrereqSeedTrait;
+
 	private string $dbdir = '';
 	private bool $hadPfb = FALSE;
 	private array $originalPfb = [];
@@ -109,51 +113,6 @@ final class SyncFeedPassDeferralTest extends TestCase
 		@rmdir($dir);
 	}
 
-	/** The minimum config pfb_global() (called at the top of sync_package_pfblockerng()) needs. */
-	private function seedSyncPrereqs(): void
-	{
-		$gen   = 'installedpackages/pfblockerng/config/0';
-		$ip    = 'installedpackages/pfblockerngipsettings/config/0';
-		$dnsbl = 'installedpackages/pfblockerngdnsblsettings/config/0';
-
-		config_set_path("{$gen}/pfb_min",        '0');
-		config_set_path("{$gen}/pfb_hour",       '0');
-		config_set_path("{$gen}/pfb_dailystart", '0');
-		config_set_path("{$gen}/skipfeed",       '0');
-		config_set_path("{$gen}/pfb_interval",   '24');
-		config_set_path("{$gen}/pfb_quiet_hours", '');
-
-		config_set_path("{$ip}/suppression",     '');
-		config_set_path("{$ip}/database_cc",     '');
-		config_set_path("{$ip}/maxmind_locale",  'en');
-		config_set_path("{$ip}/asn_reporting",   'disabled');
-		config_set_path("{$ip}/asn_token",       '');
-		config_set_path("{$ip}/maxmind_account", '');
-		config_set_path("{$ip}/maxmind_key",     '');
-
-		config_set_path('installedpackages/pfblockerngglobal/pfbextdns', '8.8.8.8');
-
-		config_set_path("{$dnsbl}/pfb_dnsvip4",     '');
-		config_set_path("{$dnsbl}/pfb_dnsvip6",     '');
-		config_set_path("{$dnsbl}/pfb_dnsport",     '8081');
-		config_set_path("{$dnsbl}/pfb_dnsport_ssl", '8443');
-		config_set_path("{$dnsbl}/top1m_enable",    '');
-		config_set_path("{$dnsbl}/pfb_cache",       '');
-		config_set_path("{$dnsbl}/pfb_py_reply",    '');
-		config_set_path("{$dnsbl}/pfb_regex",       '');
-		config_set_path("{$dnsbl}/pfb_regex_list",  '');
-		config_set_path("{$dnsbl}/pfb_cname",       '');
-		config_set_path("{$dnsbl}/tld_allow",       '');
-		config_set_path("{$dnsbl}/pfb_py_nolog",    '');
-		config_set_path("{$dnsbl}/pfb_noaaaa",      '');
-		config_set_path("{$dnsbl}/pfb_noaaaa_list", '');
-		config_set_path("{$dnsbl}/pfb_gp",          '');
-		config_set_path("{$dnsbl}/pfb_gp_bypass_list", '');
-
-		if (!isset($GLOBALS['g']['unbound_chroot_path'])) {
-			$GLOBALS['g']['unbound_chroot_path'] = '/var/unbound';
-		}
-	}
 
 	/** Hold the feed-pass lock as ANOTHER process would -- a second, independent fd. */
 	private function holdLockAsAnotherProcess(): void
